backend/requirements.txt
pandas
numpy
fastapi
uvicorn
python-dotenv
pydantic
scikit-learn
httpx
